Komolika/requirements.txt
# Core dependencies for Image Caption Generation
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.21.0
Pillow>=8.0.0
matplotlib>=3.5.0

# Web framework dependencies
Flask>=2.0.0
Werkzeug>=2.0.0

# Optional: For better progress bars
tqdm>=4.60.0
